rate changed to hz

+++ projects/hw/ob1203/sensor_lib/Inc/f103.h
@@ -62,7 +62,7 @@ void ob1203_Delay_ms(uint32_t ms);
 void ob1203_send_results(uint32_t ppg, unsigned char channel_num);
 void BMX_send_result(int16_t* acc, int16_t* giro, int16_t* mag);
 void ob1203_send_preambula(void);
-void ob1203_send_info(uint8_t rate);
+void ob1203_send_info(uint16_t rate);
 void bmx055_send_info(uint8_t rate);
